+++ History.md
@@ -1,3 +1,9 @@
+## main
+
+* Breaking changes:
+  * Digest authentication support, deprecated in 2.0.0, has been
+    removed (Jeremy Evans #307)
+
 ## 2.0.0 / 2022-06-24
 
 * Breaking changes: